Use separate conditionals for directory presence and file presence.
Previously redirect entries wouldn't be created if the directory
existed, but this broke for directories that had nested articles. The
updated conditionals test for the need to create the directory and the
existence of an index.html file separately to both avoid clobbering
existing content and create redirects from directories with nested
content.

+++ lib/nanoc-redirector.rb
@@ -17,10 +17,8 @@ module NanocRedirector
         redirects.each do |redirect|
           content = NanocRedirector.redirect_template(dest)
           dir = "output/#{redirect}"
-          unless File.directory?(dir)
-            FileUtils.mkdir_p(dir)
-            File.write("#{dir}/index.html", content)
-          end
+          FileUtils.mkdir_p(dir) unless File.directory?(dir)
+          File.write("#{dir}/index.html", content) unless File.exist? "#{dir}/index.html"
         end
       end
     end
